fix handling of 0 arguments

+++ bin/getsong.ls
@@ -5,17 +5,23 @@ require! {
 
 {exec, mkdir, ls, rm, mv} = require 'shelljs'
 
-query = process.argv[2]
-if fs.existsSync('getsongtmp')
+main = ->
+  query = process.argv[2]
+  if not query?
+    console.log 'need to provide a search query or url'
+    return
+  if fs.existsSync('getsongtmp')
+    rm '-r', 'getsongtmp'
+  mkdir 'getsongtmp'
+  process.chdir 'getsongtmp'
+  exec "you-get --itag=18 '#{query}'"
+  outfile = ls('*.mp4').stdout.trim()
+  mv outfile, 'tmp.mp4'
+  exec "ffmpeg -i tmp.mp4 -acodec copy -vn audio.m4a"
+  newname = outfile.substr(0, outfile.length - 4) + '.m4a'
+  mv 'audio.m4a', '../' + newname
+  process.chdir '..'
   rm '-r', 'getsongtmp'
-mkdir 'getsongtmp'
-process.chdir 'getsongtmp'
-exec "you-get --itag=18 '#{query}'"
-outfile = ls('*.mp4').stdout.trim()
-mv outfile, 'tmp.mp4'
-exec "ffmpeg -i tmp.mp4 -acodec copy -vn audio.m4a"
-newname = outfile.substr(0, outfile.length - 4) + '.m4a'
-mv 'audio.m4a', '../' + newname
-process.chdir '..'
-rm '-r', 'getsongtmp'
-console.log newname
+  console.log newname
+
+main()
